% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.r
\name{addDAResults}
\alias{addDAResults}
\title{Adding Differential Abundancce Results to C3NA Object}
\usage{
addDAResults(
  twoPhenoC3NAObj = twoPhenoC3NAObj,
  DAResults = NULL,
  verbose = TRUE
)
}
\arguments{
\item{twoPhenoC3NAObj}{(Required) C3NA objects after running comparePhenotypes()}

\item{DAResults}{(Required) Differential abundance results, should have a 'TaxaName' column
with any number of DA methods results. To ensure the data are selected correctly, only the
logical (TRUE/FALSE) columns will be included. Please ensure the columns are reflect the correct
type.}

\item{verbose}{(Optional) If TRUE, the function will display how many number of taxa that
shared or unique from the DA and C3NA pipeline.}
}
\description{
Adding the DA results (run separately from C3NA required pipeline) to
C3NA objects.
}
