% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.r
\docType{package}
\name{C3NA-package}
\alias{C3NA-package}
\title{Correlation and Consensus-Based Cross-Taxonmony Network Analysis}
\description{
Correlation and Consensus-based Cross-taxonomy Network Analysis (C3NA) is a
user-friendly tool for investigating compositional microbial sequencing data
to identify and compare co-occurrence patterns across different taxonomic
levels between two phenotypes.
}
\author{
Kuncheng Song
}
\keyword{package}
