% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleRelated.r
\name{comparePhenotypes}
\alias{comparePhenotypes}
\title{Obtain the Module Information from Consensus-Based Investigation}
\usage{
comparePhenotypes(
  C3NAObj_Comparison = C3NAObj1,
  C3NAObj_Reference = C3NAObj2,
  corCutoff = 0.2,
  fdr = 0.05,
  unusefulTaxa = NA,
  verbose = 0,
  nBootstrap = 300
)
}
\arguments{
\item{C3NAObj_Comparison}{(Required) The C3NAObj with a single phenotype as Disease.}

\item{C3NAObj_Reference}{(Required) The C3NAObj with a single phenotype as Control.}

\item{corCutoff}{(Required) A positive correlation cut-off point, default is 0.2.}

\item{fdr}{(Required) BH-Adjusted p-value, default is 0.05.}

\item{unusefulTaxa}{(Optional) A list of taxa names that could not be interpreated biologically, e.g. human_gut, ambiguous_taxa}

\item{verbose}{(Optinal) Verbose for the module preservation analysis, the higher the more detailed output.}

\item{nBootstrap}{(Optional) Number of bootstrap for the module preservation.}
}
\description{
Extract the module information based on the consensus results.
}
