% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMethods.R
\name{compareTwoPhenoShiny}
\alias{compareTwoPhenoShiny}
\title{Shiny Application for C3NAobj Comparison}
\usage{
compareTwoPhenoShiny(C3NAObj)
}
\arguments{
\item{C3NAObj}{(Required) The C3NAObj with Compared Phenotypes}
}
\description{
This function calls the interactive two phenotypes after running the
comparePhenotypes() step.
}
