% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{CRC_Phyloseq}
\alias{CRC_Phyloseq}
\title{A Phyloseq Object of Colorectal Cancer 16S rRNA Data}
\format{
127 Cancer and 134 Control Patients Data.
}
\usage{
data(CRC_Phyloseq)
}
\description{
The demo file includes part of the 16S rRNA processed data from Baxter et al., with 127 Cancer and 134 Control subjects.
}
\examples{
data(CRC_Phyloseq)
}
\keyword{datasets}
