% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.r
\name{extractResults}
\alias{extractResults}
\title{Extract C3NA Results}
\usage{
extractResults(twoPhenoC3NAObj = twoPhenoC3NAObj)
}
\arguments{
\item{twoPhenoC3NAObj}{C3NA objects after running comparePhenotypes()}
}
\description{
The C3NAObj after comparePhenotypes() command is quiet large due to storing accessory data for Shiny plotting.
This command is developed to extract the following information: 1. Taxa-Taxa Correlations that passed pre-set
BH-adjusted p-value cut-off, here we did NOT filter by any correlation threshold.
2. All the node data for both phenotypes, it will contain the differential abundance and C3NA
results if avaliable. This can then be used for other network plotting software or analysis.
}
