% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influentialTaxa.r
\name{getInfluentialTaxa}
\alias{getInfluentialTaxa}
\title{Adding Influential Taxa Evaluation to the C3NA Object}
\usage{
getInfluentialTaxa(C3NAObj = C3NAObj, C3NAFDR = 0.05)
}
\arguments{
\item{C3NAObj}{(Required) The C3NAObj after the comparePhenotypes() function}

\item{C3NAFDR}{(Optional) This allows to change the BH-Adjuted p-value for the paired
Wilcoxon test for degree, betweenness and transitivity}
}
\description{
Using Paired Wilcoxon Test to evaluate each taxa based on the degree,
betweenness and transitivity changes.
}
