% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleRelated.r
\name{getOptMods}
\alias{getOptMods}
\title{Obtain the Module Information from Consensus-Based Investigation}
\usage{
getOptMods(C3NAObj = C3NAObj, selectedPatterns = NA, nModules = NA)
}
\arguments{
\item{C3NAObj}{(Required) The C3NAObj with a single phenotype and after the initiateC3NA function.}

\item{selectedPatterns}{(Required) The selected unique patterns identified in the shiny application.}

\item{nModules}{(Required) The optimal number of clusters identified by the shiny application.}
}
\description{
Extract the module information based on the consensus results.
}
