% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleRelated.r
\name{getStackedTaxaMatrix}
\alias{getStackedTaxaMatrix}
\title{Obtained the stacked-taxa table}
\usage{
getStackedTaxaMatrix(
  prevTrh = 0.1,
  phyloseqObj = phyloseqObj,
  nMinTotalCount = 1000,
  phenotype = NA
)
}
\arguments{
\item{prevTrh}{(Required) Prevalence threshold of the samples, which is a number between 0 and 1. E.g., the default 0.1 represents 10\% of the samples need to have given taxa.}

\item{phyloseqObj}{(Required) Phyloseq \linkS4class{phyloseq} object. This should first undergo validatePhyloseq to ensure the diagnosis column are present.}

\item{nMinTotalCount}{(Required) The Minimal number of reads per sample. Default: 1,000.}

\item{phenotype}{(Required) The desired phenotype that present under the diagnosis column in the metadata from phyloseqObj}
}
\value{
countMatrix
}
\description{
Generate the stacked-taxa tabel for other correlation methods calculation outside C3NA.
}
\examples{
data(CRC_Phyloseq)
curPhyloseq = validatePhloseq(phyloseqObj = CRC_Phyloseq)

# These steps are commented out due to time consuming step. The post sparcc correlation data will be 
# to avoid the step
phyloseq_Cancer = phyloseq::subset_samples(physeq = CRC_Phyloseq, diagnosis == "Cancer")
stackedTaxaMatrix = getStackedTaxaMatrix(phyloseqObj = phyloseq_Cancer, phenotype = "Cancer")

# Correlation method
testCorMatrix = cor(t(stackedTaxaMatrix))
}
