% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleRelated.r
\name{initiateC3NA}
\alias{initiateC3NA}
\title{Initiate the C3NA object with Correlation Calculation and Topology Overlap Matrix}
\usage{
initiateC3NA(
  phyloseqObj = phyloseqObj,
  prevTrh = 0.1,
  nCPUs = 1,
  nBootstrap = 1000,
  nMinTotalCount = 1000,
  phenotype = NA,
  minModuleSize = 3,
  maxModuleSize = 40,
  seed = 100
)
}
\arguments{
\item{phyloseqObj}{(Required) Phyloseq \linkS4class{phyloseq} object. This should first undergo validatePhyloseq to ensure the diagnosis column are present.}

\item{prevTrh}{(Required) Prevalence threshold of the samples, which is a number between 0 and 1. E.g., the default 0.1 represents 10\% of the samples need to have given taxa.}

\item{nCPUs}{(Optional) Parallel computation for the sparccboot function. Default: 1.}

\item{nBootstrap}{(Required) Number of bootstrap for the sparcc command. Warning: this is a very computational step.}

\item{nMinTotalCount}{(Required) The Minimal number of reads per sample. Default: 1,000.}

\item{phenotype}{(Required) The desired phenotype that present under the diagnosis column in the metadata from phyloseqObj}

\item{minModuleSize}{(Optional) The lowest number for the minimal size for each cluster. Default: 3.}

\item{maxModuleSize}{(Optional) The highest number for the minimal size for each cluster. Default: 40.}

\item{seed}{(Optional)}
}
\value{
A list object of C3NA Objects
}
\description{
Generate the initiate C3NA object, including the count matrix, the taxonomic table, and the multi-taxa stacked count matrix for the C3NA analysis. This step also includes the sparcc correlation calculation. Note: sparcc correlation can be extremely computationally expensive. The results will be summarized along with the signed network calcualted from the topology overlap matrix. Finally, different clusters of taxa will be calculated based on a minimal module size range from 3 to 40 (default).
}
\examples{
data(CRC_Phyloseq)
curPhyloseq = validatePhloseq(phyloseqObj = CRC_Phyloseq)

# These steps are commented out due to time consuming step. The post sparcc correlation data will be 
# to avoid the step
phyloseq_Normal = phyloseq::subset_samples(physeq = CRC_Phyloseq, diagnosis == "Normal")
# C3NAObj_Normal = initiateC3NA(phyloseqObj = phyloseq_Normal, prevTrh = 0.1, 
#                               nCPUs = 10, nBootstrap = 1000, 
#                               nMinTotalCount = 1000, phenotype = "Normal", 
#                               seed = 100)

phyloseq_Cancer = phyloseq::subset_samples(physeq = CRC_Phyloseq, diagnosis == "Cancer")
# C3NAObj_Cancer = initiateC3NA(phyloseqObj = phyloseq_Cancer, prevTrh = 0.1, 
#                               nCPUs = 10, nBootstrap = 1000, 
#                               nMinTotalCount = 1000, phenotype = "Cancer", 
#                               seed = 100)

}
