% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMethods.R
\name{moduleEvals}
\alias{moduleEvals}
\title{Shiny Application for Optimal Module Evaluations}
\usage{
moduleEvals(C3NAObj = C3NAObj, colorSeed = 9)
}
\arguments{
\item{C3NAObj}{(Required) The C3NAObj with a single phenotype and after the initiateC3NA function.}

\item{colorSeed}{(Optional) The seed set for reproducible random colors for the bar plot fill. Default: 9.}
}
\description{
Shiny application that includes two panels. On the first panel, the user can conduct a step-by-step investigation of the consensus-based module selection, and the results will be summarized into an R command in which the user copies and pastes to the R console. The second panel includes a download window for extracting the module pattern, silhouette plot, consensus, and correlation plots.
}
