% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.r
\name{validatePhloseq}
\alias{validatePhloseq}
\title{Verify Phyloseq Objects before Running the C3NA Pipeline}
\usage{
validatePhloseq(phyloseqObj = phyloseqObj, diagColName = NA)
}
\arguments{
\item{phyloseqObj}{(Required) A phyloseq object.}

\item{diagColName}{(Optional) Identify the dianogis/phenotype column, and create a 'diagnosis' column
the column was named differently.}
}
\value{
A phyloseq object
}
\description{
This validation serve to ensure the data are prepared is ready for the C3NA.
\enumerate{
\item Ensure the OTUs/ASVs table, Taxa table and sample data with a diagnosis column present.
\item The user can provide the correct diagnosis column required for the C3NA phenotype data processing.
}
}
\examples{
data(CRC_Phyloseq)
curPhyloseq = validatePhloseq(phyloseqObj = CRC_Phyloseq)

}
